Add MQTT command to resend autodiscovery
Taken from here: https://github.com/EliasKotlyar/Xiaomi-Dafang-Hacks/pull/1051

In some cases (e.g., power failure), it is possible for the camera to
boot before the MQTT server is running. In this case, the autodiscovery
message will be lost.

To address this, this commit adds a topic called $LOCATION/set which
accepts an argument of 'announce'. This will trigger resending of the
autodiscovery message. It does not use the standard path of
$LOCATION/$DEVICE_NAME, because it is intended for all Xiaomi-Dafang-Hacks
cameras.

To use this, once your MQTT server is up, publish a message to
$LOCATION/set with the payload 'announce'. In Home Assistant, this can be done
with the following automation:

- id: '0000'
    alias: Dafang Announce on startup
    trigger:
    - event: start
      platform: homeassistant
    action:
      service: mqtt.publish
      data:
        payload: announce
        topic: myhome/set