Add 1664x936 resolution (multiply of 8).
Some NVR-software lacks the support of input resolutions which is not a multiple of 8 (f.e. MotionEye based on Motion). It's possible to use any nearest resolution (1600x904 for 1600x900 in this case), but it has a noticeable performance impact (due rescaling for preview or non-direct recording), so better way is always using the input resolution. 1664x936 has as pretty good performance as 1600x900 for this camera, the right aspect ratio, and is a multiple of 8.